+++ src/Client/ClientCore/Systems/Input/InGameKeyboardShortcuts.cs
@@ -44,6 +44,7 @@ public class InGameKeyboardShortcuts
         // For now the shortcuts are hardcoded.
         shortcutTable[SDL.SDL_Keycode.SDLK_ESCAPE] = () => Toggle(ClientStateFlag.ShowInGameMenu);
         shortcutTable[SDL.SDL_Keycode.SDLK_F2] = () => Toggle(ClientStateFlag.ShowPlayerDebug);
+        shortcutTable[SDL.SDL_Keycode.SDLK_F3] = () => Toggle(ClientStateFlag.ShowEntityDebug);
     }
 
     /// <summary>